Refactor interpreter interfaces and implementations for clarity.
