+++ interpreter/decorators.go
@@ -21,6 +21,10 @@ import (
 	"github.com/google/cel-go/common/types/traits"
 )
 
+// InterpretableDecorator is a functional interface for decorating or replacing
+// Interpretable expression nodes at construction time.
+type InterpretableDecorator func(Interpretable) (Interpretable, error)
+
 // evalObserver is a functional interface that accepts an expression id and an observed value.
 type evalObserver func(int64, ref.Val)
 
@@ -170,199 +174,4 @@ func maybeOptimizeSetMembership(i Interpretable, inlist *evalBinary) (Interpreta
 		argTypeName: typ.TypeName(),
 		valueSet:    valueSet,
 	}, nil
-}
-
-// evalSetMembership is an Interpretable implementation which tests whether an input value
-// exists within the set of map keys used to model a set.
-type evalSetMembership struct {
-	inst        Interpretable
-	arg         Interpretable
-	argTypeName string
-	valueSet    map[ref.Val]ref.Val
-}
-
-// ID implements the Interpretable interface method.
-func (e *evalSetMembership) ID() int64 {
-	return e.inst.ID()
-}
-
-// Eval implements the Interpretable interface method.
-func (e *evalSetMembership) Eval(ctx Activation) ref.Val {
-	val := e.arg.Eval(ctx)
-	if val.Type().TypeName() != e.argTypeName {
-		return types.ValOrErr(val, "no such overload")
-	}
-	if ret, found := e.valueSet[val]; found {
-		return ret
-	}
-	return types.False
-}
-
-// evalWatch is an Interpretable implementation that wraps the execution of a given
-// expression so that it may observe the computed value and send it to an observer.
-type evalWatch struct {
-	inst     Interpretable
-	observer evalObserver
-}
-
-// ID implements the Interpretable interface method.
-func (e *evalWatch) ID() int64 {
-	return e.inst.ID()
-}
-
-// Eval implements the Interpretable interface method.
-func (e *evalWatch) Eval(ctx Activation) ref.Val {
-	val := e.inst.Eval(ctx)
-	e.observer(e.inst.ID(), val)
-	return val
-}
-
-// evalExhaustiveOr is just like evalOr, but does not short-circuit argument evaluation.
-type evalExhaustiveOr struct {
-	id  int64
-	lhs Interpretable
-	rhs Interpretable
-}
-
-// ID implements the Interpretable interface method.
-func (or *evalExhaustiveOr) ID() int64 {
-	return or.id
-}
-
-// Eval implements the Interpretable interface method.
-func (or *evalExhaustiveOr) Eval(ctx Activation) ref.Val {
-	lVal := or.lhs.Eval(ctx)
-	rVal := or.rhs.Eval(ctx)
-	lBool, lok := lVal.(types.Bool)
-	if lok && lBool == types.True {
-		return types.True
-	}
-	rBool, rok := rVal.(types.Bool)
-	if rok && rBool == types.True {
-		return types.True
-	}
-	if lok && rok {
-		return types.False
-	}
-	if types.IsUnknown(lVal) {
-		return lVal
-	}
-	if types.IsUnknown(rVal) {
-		return rVal
-	}
-	return types.ValOrErr(lVal, "no such overload")
-}
-
-// evalExhaustiveAnd is just like evalAnd, but does not short-circuit argument evaluation.
-type evalExhaustiveAnd struct {
-	id  int64
-	lhs Interpretable
-	rhs Interpretable
-}
-
-// ID implements the Interpretable interface method.
-func (and *evalExhaustiveAnd) ID() int64 {
-	return and.id
-}
-
-// Eval implements the Interpretable interface method.
-func (and *evalExhaustiveAnd) Eval(ctx Activation) ref.Val {
-	lVal := and.lhs.Eval(ctx)
-	rVal := and.rhs.Eval(ctx)
-	lBool, lok := lVal.(types.Bool)
-	if lok && lBool == types.False {
-		return types.False
-	}
-	rBool, rok := rVal.(types.Bool)
-	if rok && rBool == types.False {
-		return types.False
-	}
-	if lok && rok {
-		return types.True
-	}
-	if types.IsUnknown(lVal) {
-		return lVal
-	}
-	if types.IsUnknown(rVal) {
-		return rVal
-	}
-	return types.ValOrErr(lVal, "no such overload")
-}
-
-// evalExhaustiveConditional is like evalConditional, but does not short-circuit argument
-// evaluation.
-type evalExhaustiveConditional struct {
-	id     int64
-	expr   Interpretable
-	truthy Interpretable
-	falsy  Interpretable
-}
-
-// ID implements the Interpretable interface method.
-func (cond *evalExhaustiveConditional) ID() int64 {
-	return cond.id
-}
-
-// Eval implements the Interpretable interface method.
-func (cond *evalExhaustiveConditional) Eval(ctx Activation) ref.Val {
-	cVal := cond.expr.Eval(ctx)
-	tVal := cond.truthy.Eval(ctx)
-	fVal := cond.falsy.Eval(ctx)
-	cBool, ok := cVal.(types.Bool)
-	if !ok {
-		return types.ValOrErr(cVal, "no such overload")
-	}
-	if cBool {
-		return tVal
-	}
-	return fVal
-}
-
-// evalExhaustiveFold is like evalFold, but does not short-circuit argument evaluation.
-type evalExhaustiveFold struct {
-	id        int64
-	accuVar   string
-	iterVar   string
-	iterRange Interpretable
-	accu      Interpretable
-	cond      Interpretable
-	step      Interpretable
-	result    Interpretable
-}
-
-// ID implements the Interpretable interface method.
-func (fold *evalExhaustiveFold) ID() int64 {
-	return fold.id
-}
-
-// Eval implements the Interpretable interface method.
-func (fold *evalExhaustiveFold) Eval(ctx Activation) ref.Val {
-	foldRange := fold.iterRange.Eval(ctx)
-	if !foldRange.Type().HasTrait(traits.IterableType) {
-		return types.ValOrErr(foldRange, "got '%T', expected iterable type", foldRange)
-	}
-	// Configure the fold activation with the accumulator initial value.
-	accuCtx := varActivationPool.Get().(*varActivation)
-	accuCtx.parent = ctx
-	accuCtx.name = fold.accuVar
-	accuCtx.val = fold.accu.Eval(ctx)
-	iterCtx := varActivationPool.Get().(*varActivation)
-	iterCtx.parent = accuCtx
-	iterCtx.name = fold.iterVar
-	it := foldRange.(traits.Iterable).Iterator()
-	for it.HasNext() == types.True {
-		// Modify the iter var in the fold activation.
-		iterCtx.val = it.Next()
-
-		// Evaluate the condition, but don't terminate the loop as this is exhaustive eval!
-		fold.cond.Eval(iterCtx)
-
-		// Evalute the evaluation step into accu var.
-		accuCtx.val = fold.step.Eval(iterCtx)
-	}
-	// Compute the result.
-	res := fold.result.Eval(accuCtx)
-	varActivationPool.Put(iterCtx)
-	varActivationPool.Put(accuCtx)
-	return res
 }
